fix bug bw wlan

+++ status.py
@@ -173,20 +173,24 @@ def getStatus():
     #print(list_file[1],"-",re['network']['eth0']['rx_bytes'],"=",recieve)
     re['network']['eth0']['send']=send
     re['network']['eth0']['recieve']=recieve
-    
+
     send_wlan=0
     recieve_wlan=0
-    list_file_wlan=[]
-    with open(filename_bw_wlan,"r+") as file:
-        list_file_wlan=file.read().replace("\n","").split(":")
-        file.seek(0)
-        file.write(str(re['network']['wlan0']['tx_bytes'])+":"+str(re['network']['wlan0']['rx_bytes']))
-        file.truncate()
-    #print(list_file_wlan)
-    send_wlan=abs(int(list_file_wlan[0])-re['network']['wlan0']['tx_bytes'])/1000
-    recieve_wlan=abs(int(list_file_wlan[1])-re['network']['wlan0']['rx_bytes'])/1000
-    #print(list_file_wlan[0],"-",re['network']['wlan0']['tx_bytes'],"=",send_wlan)
-    #print(list_file_wlan[1],"-",re['network']['wlan0']['rx_bytes'],"=",recieve_wlan)
+
+    if not "Device not found" in subprocess.getoutput("ifconfig wlan0"):
+            
+        list_file_wlan=[]
+        with open(filename_bw_wlan,"r+") as file:
+            list_file_wlan=file.read().replace("\n","").split(":")
+            file.seek(0)
+            file.write(str(re['network']['wlan0']['tx_bytes'])+":"+str(re['network']['wlan0']['rx_bytes']))
+            file.truncate()
+        #print(list_file_wlan)
+        print(type(re['network']['wlan0']['tx_bytes']))
+        send_wlan=abs(int(list_file_wlan[0])-int(re['network']['wlan0']['tx_bytes']))/1000
+        recieve_wlan=abs(int(list_file_wlan[1])-int(re['network']['wlan0']['rx_bytes']))/1000
+        #print(list_file_wlan[0],"-",re['network']['wlan0']['tx_bytes'],"=",send_wlan)
+        #print(list_file_wlan[1],"-",re['network']['wlan0']['rx_bytes'],"=",recieve_wlan)
     re['network']['wlan0']['send']=send_wlan
     re['network']['wlan0']['recieve']=recieve_wlan
     #re['result']="-"
